backend/requirements.txt
fastapi
uvicorn
docker
websockets
python-multipart
sqlalchemy
alembic
bcrypt
python-jose[cryptography]
pydantic[email]